Remove Event in favor of ReportEvent
