--- src/ContentSecurityPolicy/Violation/Event.php
@@ -1,34 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Nelmio SecurityBundle.
- *
- * (c) Nelmio <hello@nelm.io>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Nelmio\SecurityBundle\ContentSecurityPolicy\Violation;
-
-use Symfony\Contracts\EventDispatcher\Event as SymfonyEvent;
-
-/**
- * @deprecated since nelmio/security-bundle 2.12, use ReportEvent instead.
- */
-class Event extends SymfonyEvent
-{
-    private Report $report;
-
-    public function __construct(Report $report)
-    {
-        $this->report = $report;
-    }
-
-    public function getReport(): Report
-    {
-        return $this->report;
-    }
-}
