--- src/ContentSecurityPolicy/Violation/Events.php
@@ -1,22 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Nelmio SecurityBundle.
- *
- * (c) Nelmio <hello@nelm.io>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Nelmio\SecurityBundle\ContentSecurityPolicy\Violation;
-
-final class Events
-{
-    /**
-     * @deprecated since nelmio/security-bundle 2.12, use `ReportEvent::class` instead.
-     */
-    public const VIOLATION_REPORT = 'csp.violation.report';
-}
