+++ src/Controller/ContentSecurityPolicyController.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Nelmio\SecurityBundle\Controller;
 
-use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Events;
 use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Exception\ExceptionInterface;
 use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Filter\Filter;
 use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Log\Logger;
@@ -50,7 +49,7 @@ final class ContentSecurityPolicyController
             return new Response('', 204);
         }
 
-        $this->dispatcher->dispatch(new ReportEvent($report), Events::VIOLATION_REPORT);
+        $this->dispatcher->dispatch(new ReportEvent($report));
 
         $this->logger->log($report);
