--- tests/ContentSecurityPolicy/Violation/ReportEventTest.php
@@ -1,29 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Nelmio SecurityBundle.
- *
- * (c) Nelmio <hello@nelm.io>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace ContentSecurityPolicy\Violation;
-
-use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Report;
-use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\ReportEvent;
-use PhpUnit\Framework\TestCase;
-
-class ReportEventTest extends TestCase
-{
-    public function testCanBeConstructed()
-    {
-        $this->assertInstanceOf(
-            ReportEvent::class,
-            new ReportEvent(new Report([]))
-        );
-    }
-}
